+++ Rocket.Unturned/U.cs
@@ -283,6 +283,7 @@ namespace Rocket.Unturned
             PlayerLife.OnTellVirus_Global += UnturnedPlayerEvents.InternalOnTellVirus;
             PlayerLife.OnTellBleeding_Global += UnturnedPlayerEvents.InternalOnTellBleeding;
             PlayerLife.OnTellBroken_Global += UnturnedPlayerEvents.InternalOnTellBroken;
+            PlayerStance.OnStanceChanged_Global += UnturnedPlayerEvents.InternalOnStanceChanged;
 
             ChatManager.onCheckPermissions += (SteamPlayer player, string text, ref bool shouldExecuteCommand, ref bool shouldList) =>
             {