+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes should be documented in this file. The format is based on [K
 ### Changed
 - Replaced `OnPlayerWear` usage of `SteamChannel.onTriggerSend` with `PlayerClothing.OnItemChanged_Global`.
 - Replaced life stat usage event of `SteamChannel.onTriggerSend` with `PlayerLife.OnTellStat_Changed`.
+- Replaced `OnPlayerUpdateStance` usage of `SteamChannel.onTriggerSend` with `PlayerStance.OnStanceChanged_Global`.
 
 ## 4.9.3.6 - 2020-10-16