+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes should be documented in this file. The format is based on [K
 - Replaced `OnPlayerUpdateExperience` usage of `SteamChannel.onTriggerSend` with `PlayerSkills.OnExperienceChanged_Global`.
 - Replaced `OnPlayerUpdateStance` usage of `SteamChannel.onTriggerSend` with `PlayerStance.OnStanceChanged_Global`.
 - Replaced `OnPlayerDead` and `OnPlayerDeath` usage of `SteamChannel.onTriggerSend` with `PlayerLife.onPlayerDied`.
+- Replaced `OnPlayerUpdateLife` and `OnPlayerRevive` usage of `SteamChannel.onTriggerSend` with `PlayerLife.OnRevived_Global`.
 
 ## 4.9.3.6 - 2020-10-16